Several Keyboard/Gamepad input improvements
This is a set of changes to make the launcher usable in
Android boxes connected to TVs and so, where it's not possible
to use the touchscreen.

The Keyboard navigation not always worked correctly before, the
gridview was sometimes using its selector and other times it wasn't
the "ENTER" key wasn't behaving properly when the gridview selector
was on, etc. These hurdles are fixed with this commit.

I added some color for the selector when focused so the selected item
is actually visible, as well was the menu/search buttons.

I also added some additional bindings for keyboard and gamepad:

* Alt+O: open options (only if not password locked)
* Alt+H: toggle showing the hidden apps (only if not password locked)
* Escape Key: same behavior as back button
* Gamepad Y: context menu of the app selected
* Gamepad A: same behavior as keyboard enter
* Gamepad B: same behavior as back button
* Gamepad Select: open the options menu (asks password if locked)
* Gamepad Left bumper/trigger: switch to previous category
* Gamepad Right bumper/trigger: switch to next category

Also, there's a "search as you type" now, so the search will
automatically open when typing on the physical keyboard.